[ADD] edi_stock: Move report document and Move cancellation report
The concept currently exists where EDI can track individual picks which
have been reported on with the `_edi_pick_report_via`. This extends that
concept with a new model `edi.move.report.document` which makes use of
`_edi_move_report_via`.
The same mechanism is used for this, however instead
of having a pick --> moves, and binding the edi doc id to the pick, we
simply have moves, and bind the edi doc id to the moves.

Added ontop of this concept is a move cancellation report, which extends
the base class slightly to generate an outbound email for any new move
cancellations against picking types which have the appropriate config.
The email template which is used is also configured on the picking type,
so it is possible to create different email templates for different pick
types, and have them all batched and sent by the same edi document.

A default mail template is included too, which has some sane defaults

An additional module edi_sale_stock has been added which extends the
functionality of this report to include information from the sale

Signed-off-by: Peter Alabaster <peter.alabaster@unipart.io>

User-story: 14633