[ADD] edi: Log edi document prepare/execute processing status
SE-1165

While a large EDI document is being prepared/executed, which may create
a large number of records taking long (in background processing etc.),
let support users view the progress logging messages.

As EDI documents are prepared/executed under a DB `savepoint`, the
progress log messages need to be committed to a related model
(`edi.document.progress`) in a separate cursor to not disturb the main
transcation that creates edi records and target records.

An odoo-core bug caused yet-to-be written edi/target records also
committed when the progress log is committed in a different cursor, it
is worked around by clearing the `self.env.all.towrite` dict (which
holds records to be committed) and `self.env.all.tocompute` (which
holds computed fields) before updating the progress record and
restoring it after. See Odoo upstream bug 65036.

Signed-off-by: Rajeesh KV <rajeesh.kv@unipart.com>