[ADD] edi: Processing granularity for received documents
SE-1165

Increase processing granularity when receiving an EDI documents so that
it can be only received, prepared or executed. EDI gateway has
`allow_process` setting that triggers `edi.transfer` with this flag.
Check the processing level defined on the document type to decided
whether the document should be prepared or processed automatically.

Signed-off-by: Rajeesh KV <rajeesh.kv@unipart.com>