[FIX] edi: Patch cursor handling for unit tests
SE-1165

Unit tests patched `commit` function for current cursor object to avoid
commits. For the progress reporting, this causes error as the
edi.document record/key is not found in the new cursor opened for
progress reporting. To avoid foreign key referencing, change the
`doc_id` field in edi.document.progress to an Integer field, yet keep
the link to the document manually.

Signed-off-by: Rajeesh KV <rajeesh.kv@unipart.com>